Added two integration tests that ensure correct functioning of controller code as we refactor it.
These tests identified three bugs that have been fixed in the code (so that the tests get correct results when run against the code).  The bugs were:

1) An incorrect use of "body" instead of "title" in the "voluptatibus"/"voluptatem" rule.

2) Filtering out only two of the three admin ids and not filtering at all for the "voluptatibus"/"voluptatem" rule.

3) Allowing a duplicate to make its way into the final list because one post matched multiple rules.

Also in the process of writing the tests worked out safer and more expressive Linq expressions that fetch the correct data from the web service.  We will eventually replace the production logic with these.

To allow test to fetch data to derive expected results, created a PostRepository class that exposes a Posts static property, allowing retrieval of all posts. This property populates the collection the first time it is invoked.   First refactoring step will be to remove Posts cache from the Controller and substitute use of this repository.  Wanted to test and fix existing code first, however, so that we are refactoring correct logic.