Added a IRetrievePostData interface and a RetrievePostData concrete class that integrates it that pulls its data from the Web Service.
Added a parameter to PostRepository's static Configure method that injects an IRetrievePostData in addition to the logging interace previously injected.

Added one integration test that instantiates a concrete instance of RetrievePostData and passes this into Configure so that PostRepository will fetch its data from the web service for this test.

Modified the other existing tests to pass a mock IRetrievePostsData whose GetAllPosts is set up to return hard-coded mock data.